third_party/WebKit/LayoutTests/http/tests/inspector/network/ping-response.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>

function sendBeacon()
{
    console.log("Beacon sent: " + navigator.sendBeacon("resources/empty.html", "foo"));
}

function test() {
  ConsoleTestRunner.addConsoleSniffer(step2);
  TestRunner.evaluateInPage('sendBeacon()');

  function step2() {
    NetworkTestRunner.networkRequests().pop().requestContent().then(step3);
  }

  function step3() {
    var request = NetworkTestRunner.networkRequests().pop();
    TestRunner.addResult('URL: ' + request.url());
    TestRunner.addResult('Finished: ' + request.finished);
    TestRunner.addResult('Cached: ' + request.cached());
    TestRunner.addResult('Method: ' + request.requestMethod);
    TestRunner.addResult('Status: ' + request.statusCode + ' ' + request.statusText);
    TestRunner.addResult('Has raw request headers: ' + (typeof request.requestHeadersText() === 'string'));
    TestRunner.addResult('Has raw response headers: ' + (typeof request.responseHeadersText === 'string'));
    TestRunner.completeTest();
  }
}
</script>
</head>
<body onload="runTest()">
<p>Test that ping request response is recorded.</p>
</body>
</html>
