third_party/WebKit/LayoutTests/http/tests/inspector/network/ping-response-expected.txt

CONSOLE MESSAGE: line 9: Beacon sent: true
Test that ping request response is recorded.

URL: http://127.0.0.1:8000/inspector/network/resources/empty.html
Finished: true
Cached: false
Method: POST
Status: 200 OK
Has raw request headers: true
Has raw response headers: true
